Axiom/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.13
filetype==1.2.0
frozenlist==1.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grandalf==0.8
greenlet==3.2.3
groq==0.28.0
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_pydantic==0.6
jupyter_client==8.6.3
jupyter_core==5.8.1
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-google-genai==2.1.5
langchain-groq==0.3.2
langchain-mcp-adapters==0.1.7
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-checkpoint-sqlite==2.0.10
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.9.4
mcp-use==1.3.0
mem0ai==0.1.108
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.2.6
openai==1.86.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
portalocker==2.10.1
posthog==4.10.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
qdrant-client==1.14.2
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlite-vec==0.1.6
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.46.2
tenacity==9.1.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
